refactor: rename mocks to mockFunctions.jsx
