+++ src/__test__/helpers/store.test.tsx
@@ -10,7 +10,7 @@ import {
 } from '../../utils/store';
 import { mockStore } from '../mocks/store';
 import { bulbasaur } from '../mocks/data';
-import { dispatchSpy } from '../mocks/mocks';
+import { dispatchSpy } from '../mocks/mockFunctions';
 import { MockProvider } from '../mocks/MockProvider';
 
 describe('store:', () => {
